requirements.txt
# data-research-agent — Python Dependencies
# Install with: uv pip install -r requirements.txt

# LangChain (pinned for compatibility) 
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-ollama==0.2.1

# Vector Store 
chromadb>=0.5.0

#  Data Platform API Clients 
kaggle>=1.6.0
huggingface-hub>=0.25.0

#  HTTP Client 
httpx>=0.27.0
requests>=2.31.0

# Data Processing 
pandas>=2.2.0
pydantic>=2.8.0

#  CLI & Terminal Output 
typer>=0.12.0
rich>=13.0.0

# Environment & Config 
python-dotenv>=1.0.0

# Database 
sqlalchemy>=2.0.0

#  Testing 
pytest>=8.0.0
pytest-asyncio>=0.23.0

#  HTML Parsing (UCI scraping fallback)
beautifulsoup4>=4.12.0